+++ src/js/index.js
@@ -1,6 +1,8 @@
-import './components/header/header';
+import './global';
+
 import './components/corsine/corsine';
 import './components/brands/brands';
 import './components/product/product';
 import './components/order/order';
+import './components/menu/menu';
