+++ src/scss/blocks/corsine.scss
@@ -10,6 +10,7 @@
         width: 500px;
         right: 0;
         top: calc(100% + 40px);
+        display: none;
     }
 
     &__header {
