+++ src/scss/blocks/header.scss
@@ -14,12 +14,12 @@
         }
     }
 
-    &__contacts {
+    &__worktime {
         display: flex;
         margin-left: -15px;
     }
 
-    &__contacts-item {
+    &__text {
         margin-left: 15px;
         padding-right: 15px;
         color: $color-secondary;
@@ -28,23 +28,30 @@
         &:not(:last-child) {
             border-right: 1px solid $color-secondary;
         }
+    }
 
-        &_phone {
-            position: relative;
-            padding-left: 31px;
+    &__phone {
+        position: relative;
+        padding-left: 31px;
+        color: $color-secondary;
+        font-size: 12px;
+        transition: color .2s;
 
-            &:before {
-                content: '';
-                position: absolute;
-                left: 0;
-                top: 50%;
-                transform: translateY(-50%);
-                background: url('../images/icons/icons.svg#phone');
-                width: 16px;
-                height: 16px;
-                display: block;
-                color: gray;
-            }
+        &:before {
+            content: '';
+            position: absolute;
+            left: 0;
+            top: 50%;
+            transform: translateY(-50%);
+            background: url('../images/icons/icons.svg#phone');
+            width: 16px;
+            height: 16px;
+            display: block;
+            color: gray;
+        }
+
+        &:hover {
+            color: #999;
         }
     }
 
@@ -87,36 +94,11 @@
         margin-right: 60px;
     }
 
-    &__search {
-        margin-left: auto;
-
-        &-toggle {
-            position: relative;
-            display: none;
-        }
-
-        &-field {
-            width: 300px;
-            padding: 10px;
-            background: #0f0f0f;
-            border: 1px solid #aaa;
-            border-radius: 6px;
-            outline: none;
-            color: #fff;
-            transition: border .2s, color .2s;
-    
-            &::placeholder {
-                color: #aaa;
-            }
-            
-            &:focus {
-                border: 1px solid #ddd;
-    
-                &::placeholder {
-                    color: #ddd;
-                }
-            }
-        }
+    &__cart-btn {
+        background: none;
+        border: none;
+        cursor: pointer;
+        outline: none;
     }
 }
 
@@ -124,44 +106,4 @@
     font-size: 36px;
     text-transform: uppercase;
     font-family: 'PostNoBillsJaffna', sans-serif;
-}
-
-.menu {
-    &__list {
-        display: flex;
-        margin: 0;
-    }
-
-    &__link {
-        text-transform: uppercase;
-        padding: 32px 20px;
-        display: block;
-        color: $color-secondary;
-        font-size: 14px;
-
-        &:hover {
-            background: #1F1F1F;
-            color: $color-primary;
-        }
-    }
-
-    &_active {
-        .toggle-menu__lines {
-            margin-bottom: 4px;
-            width: 0;
-            transition: width 0s;
-        
-            &:before {
-                top: 0;
-                border-radius: 2px;
-                transform: rotate(-45deg);
-            }
-        
-            &:after {
-                border-radius: 2px;
-                top: 0;
-                transform: rotate(45deg);
-            }
-        }
-    }
 }
