+++ src/scss/blocks/main-banner.scss
@@ -1,6 +1,6 @@
 .main-banner {
     background-color: #0f0f0f;
-    background-image: url('../images/header.jpg');
+    background-image: url('../images/header2.jpg');
     background-position: center center;
     background-repeat: no-repeat;
     background-size: cover;
@@ -11,6 +11,22 @@
     justify-content: center;
     align-items: center;
     min-height: 600px;
+    position: relative;
+    padding-top: 45px;
+
+    &:before {
+        content: '';
+        position: absolute;
+        left: 0;
+        top: 0;
+        width: 100%;
+        height: 100%;
+        background: rgba(0, 0, 0, .7);
+    }
+
+    &__content {
+        position: relative;
+    }
     
     &__main-inf {
         height: 320px;
@@ -50,5 +66,6 @@
         line-height: 24px;    
         text-align: center;
         padding: 0 10px;
+        text-shadow: 2px 2px 7px #000;
     }
 }
