+++ src/scss/blocks/menu.scss
@@ -1,0 +1,96 @@
+.menu {
+    &__toggle {
+        $toggleColor: #fff;
+        $toggleLineHeight: 4px;
+        $toggleLineWidth: 30px;
+        $toggleBorderRadius: 2px;
+        $toggleTransition: .3s;
+        
+        display: none;
+        height: calc(#{$toggleLineHeight} * 5);
+        position: relative;
+        width: $toggleLineWidth;
+        cursor: pointer;
+        background: none;
+        border: none;
+      
+        &:focus {
+            outline: none;
+        } 
+      
+        &-lines {
+            position: absolute;
+            top: 50%;
+            right: 0;
+            transform: translateY(-50%);
+            height: $toggleLineHeight;
+            width: $toggleLineWidth;
+            background: $toggleColor;
+            border-radius: $toggleBorderRadius;
+            transition: width $toggleTransition;
+      
+            &:before {
+                content: '';
+                position: absolute;
+                top: -200%;
+                right: 0;
+                background: $toggleColor;
+                height: $toggleLineHeight;
+                width: $toggleLineWidth;
+                border-radius: $toggleBorderRadius;
+                transition: transform $toggleTransition;
+                transform: rotate(0deg);
+            }
+      
+            &:after {
+                content: '';
+                position: absolute;
+                top: 200%;
+                right: 0;
+                background: $toggleColor;
+                height: $toggleLineHeight;
+                width: $toggleLineWidth;
+                border-radius: $toggleBorderRadius;
+                transition: transform $toggleTransition;
+            }
+        }
+    }
+
+    &__list {
+        display: flex;
+        margin: 0;
+    }
+
+    &__link {
+        text-transform: uppercase;
+        padding: 23px 20px;
+        display: block;
+        color: $color-secondary;
+        font-size: 14px;
+
+        &:hover {
+            background: #1F1F1F;
+            color: $color-primary;
+        }
+    }
+
+    &_active {
+        .menu__toggle-lines {
+            margin-bottom: 4px;
+            width: 0;
+            transition: width 0s;
+        
+            &:before {
+                top: 0;
+                border-radius: 2px;
+                transform: rotate(-45deg);
+            }
+        
+            &:after {
+                border-radius: 2px;
+                top: 0;
+                transform: rotate(45deg);
+            }
+        }
+    }
+}
