+++ src/scss/media/main.scss
@@ -4,3 +4,4 @@
 @import 'news';
 @import 'catalog';
 @import 'footer';
+@import 'menu';
