+++ src/scss/media/menu.scss
@@ -1,0 +1,34 @@
+.menu {
+    @media screen and (max-width: 980px) {
+        & {
+            order: 5;
+            margin: 0;
+            position: relative;
+        }
+        
+        &__list {
+            position: fixed;
+            right: 0;
+            margin-top: 9px;
+            transform: translateX(100%);
+            background: $color-bg;
+            transition: transform .3s;
+            display: block;
+        }
+    
+        &__link {
+            padding: 15px 30px;
+            color: #f1f1f1;
+        }
+    
+        &__toggle {
+            display: block;
+        }
+    
+        &_active {
+            .menu__list {
+                transform: translateX(0);
+            }
+        }
+    }
+}
