+++ src/scss/styles.scss
@@ -15,6 +15,7 @@
 @import 'blocks/footer';
 @import 'blocks/corsine';
 @import 'blocks/order';
+@import 'blocks/menu';
 
 @import 'other';
